requirements.txt
numpy
pandas
scikit-learn
xgboost
matplotlib
seaborn
tqdm
opencv-python
mediapipe
pyyaml
torch
torchvision
torchaudio
joblib
